+++ app-backend/db/src/main/resources/migrations/V76__add_previous_session_to_sessions.sql
@@ -1,0 +1,4 @@
+ALTER TABLE
+    task_sessions
+ADD
+    COLUMN previous_session_id uuid REFERENCES task_sessions (id);